fix: use react router for page navigation
