+++ src/App.tsx
@@ -2,9 +2,32 @@ import "./App.css";
 import { IntlProvider } from "react-intl";
 import { getTranslations } from "./lib/translations";
 import { Header } from "./components/Header";
-import { BodyComponent } from "./components/Body";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { FormattedMessage } from "react-intl";
+import { useEffect, useState } from "react";
+import trustChainList from "./assets/trustChainList.json";
+import axios from "axios";
+import { UrlInput } from "./components/UrlInput";
+import { AnchorList } from "./components/AnchorList";
+import { RestoreView } from "./components/RestoreView";
+import { GraphView } from "./components/GraphView";
+import bodyStyle from "./css/BodyComponent.module.css";
 
-function App() {
+const App = () => {
+  const [corsEnabled, setCorsEnabled] = useState(false);
+
+  useEffect(() => {
+    if (trustChainList.length) {
+      const testUrl =
+        `${import.meta.env.VITE_CORS_PROXY}` ||
+        "" + trustChainList[0].url + "/.well-known/openid-federation";
+
+      axios.get(testUrl).catch((e) => {
+        if (e.request.status === 0) setCorsEnabled(true);
+      });
+    }
+  });
+
   return (
     <div className="App">
       <IntlProvider
@@ -12,8 +35,36 @@ function App() {
         defaultLocale="en-EN"
         messages={getTranslations(navigator.language)}
       >
-        <Header />
-        <BodyComponent />
+        <BrowserRouter basename={import.meta.env.VITE_BASE_PATH}>
+          <Header />
+          {corsEnabled && (
+            <div
+              className="alert alert-warning"
+              role="alert"
+              style={{ fontSize: "14px" }}
+            >
+              <FormattedMessage id="cors_warning" />
+              <a
+                href={`${import.meta.env.VITE_CORS_DOCS_URL}` || "/"}
+                className="alert-link"
+                style={{ marginLeft: "10px" }}
+              >
+                <FormattedMessage id="read_more" />
+              </a>
+            </div>
+          )}
+
+          <div className={bodyStyle.bodyContainer}>
+            <Routes>
+              <Route path="/" Component={UrlInput} />
+              <Route path="/insertUrl" Component={UrlInput} />
+              <Route path="/listUrl" Component={AnchorList} />
+              <Route path="/restoreSession" Component={RestoreView} />
+              <Route path="/insertEntityUrl" Component={UrlInput} />
+              <Route path="/graphView" Component={GraphView} />
+            </Routes>
+          </div>
+        </BrowserRouter>
         <canvas id="canvas" style={{ display: "none" }}></canvas>
       </IntlProvider>
     </div>
